+++ Kuripawarudo_Linux/sources/block.cpp
@@ -22,33 +22,33 @@ Block::Block (){}
  * @param type Le type de bloc.
  * @param rocheType 1, 2 ou 3
  */
-Block::Block (unsigned int type, unsigned int rocheType)
+Block::Block (uint8_t type, uint8_t rocheType)
 {
     switch (type)
     {
-    case TERRE:
-        // Création bloc de terre.
-        this->type = type ;
-        soliditee   = 3 ;
-        
-        break ;
-    case ROCHE:
-        // Création bloc de roche.
-        this->type      = type ;
-        this->rocheType = rocheType ;
-        switch (rocheType)
-        {
-            case 1: soliditee = 6  ; break ;
-            case 2: soliditee = 12 ; break ;
-            case 3: soliditee = 24 ; break ;
-        }
-        
-        break ;
-    case 3:
-        // Création bloc de minerai.
-        this->type        = type ;
-        minerai_quantitee = rand() % 500 + 1 ;
-        break ;
+        case TERRE:
+            // Création bloc de terre.
+            this->type = type ;
+            soliditee   = 3 ;
+            
+            break ;
+        case ROCHE:
+            // Création bloc de roche.
+            this->type      = type ;
+            this->rocheType = rocheType ;
+            switch (rocheType)
+            {
+                case 1: soliditee = 6  ; break ;
+                case 2: soliditee = 12 ; break ;
+                case 3: soliditee = 24 ; break ;
+            }
+            
+            break ;
+        case 3:
+            // Création bloc de minerai.
+            this->type        = type ;
+            minerai_quantitee = rand() % 500 + 1 ;
+            break ;
     }
 
 }
@@ -68,7 +68,7 @@ Block::~Block (){}
  * 
  * @return Le type de bloc.
  */
-unsigned int Block::getType ()
+uint8_t Block::getType ()
 {
     return type ;
 }
@@ -78,7 +78,7 @@ unsigned int Block::getType ()
  * 
  * @return La valeur de la variable rocheType.
  */
-unsigned int Block::getRocheType ()
+uint8_t Block::getRocheType ()
 {
     return rocheType ;
 }
